Flask app and database connection working
The application is now set up with Flask, simple sign up and log in
functionality have been added which allow for a user to sign up and log
in, including some authentication and persistence due to the SQLAlchemy
database. The overall structure of the applictaion follows recommended
standards and the README has been updated to include a reference used
for development assistance.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>