Added page for clustering result and hid views
The group page will be used to display the cluster that the user falls
into based on their answers to a quiz. That view is currently hidden
until the user is signed in atleast. I have also added a gitignore to
remove files that should not be versioned in git. I have also added the
log out functionality to the nav bar so that it can be accessed.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>