Quiz data turned into dict and stored properly
The quiz data was causing issues and not saving to the database as the
survey view was never becoming equal to post, causing the questions and
answers never to be saved to the database. Now this is fixed and data is
stored as a dict using json. The username is also now stored alongside
the questions and answers for ease of identifying the users for matching
later.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>