Added generated dataset, kmodes and feed
Now that I have created the code seperately which generates a larger
dataset of user responses based on real world answers and the trends
acknowledged in that data, a larger dataset is now obtained. As well as
this, I have modified the clustering code to use the K-Modes clustering
algorithm instead of K-Means as K-Modes is more appropriate for
categoric data. Finally, I also created an Instagram account for the
website and populated it with posts before using Taggbox as a platform
for displaying the posts as a live feed by embedding the chosen content.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>