Cluster changes, favicon and error handling
I have now modified the code so that there are more clusters, this is
more so that I can observe changes to the cluster centroids. I have also
added a random seed which allows for the cluster attributes to remain
consistent each time the code is run. I have also added error handling
for when a user leaves a radio button option blank for a question which
is not supported. Now the issue is flashed on screen so that the user
cannot progress until they have answered every question.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>