Optimal k value added and cluster descriptions
I created a .ipynb file in jupyter lab to plot an elbow method graph
which allowed me to identify 9 as the best number of clusters for my
data. I then printed out the cluster centroids and summarised the
characteristics and trends that I acknowledged for each cluster so now a
description of each cluster is printed out depending on the category
that the user was placed in.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>