Compatibility assessment started
In this added code, now the match.html page has an input field and button
where the user can enter the username of the user they'd like to assess
their compatibility with and click on the button to do so. In the
background, the code checks if the username entered exists in the
database as a user, if it does then the application retrieves the latest
quiz responses that the currently logged in user gave and the selected
user who's username was inputted. From there, the application once again
makes sure that these answers exist and the responses are isolated
before they're passed into the 'calculate_score' function which
currently is just hardcoded to return the score as 100. This will change
once I add the correct functionality here.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>