Added score calculation
Now the compatibility score is calculated by taking each answer for both
users one by one, the amount of matches are divided by the amount of
questions, multiplied by 100 to make it into a percentage and rounded. I
hope to later enhance this further since Hamming distance function was
no longer suitable as the data handled is categoric now due to the use
of KModes instead of KMeans.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>