Graphed most popular quiz answers
Modified the code so that now the user is presented with their answers
to the survey questions and the distribution of popularity for each
answer is graphed for each question. The graphs are stored in a memory
buffer, the images are base64 formatted so that they are embedded on the
html page.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>