Footer added to all primary pages
The footer is created now to improve the overall look and completeness
of the application and to help with navigation and information being
accessed. It has been added to all pages aside from the login and sign
up pages for now since most of the links in the footer are not
accessible to users who are not logged in.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>