Users can now opt-out of public assessment
I have now implemented a suggestion that was made at my first FYP demo
which was to allow the users to opt-out of having their quiz responses
used during compatibility assessment in the case of another user
possibly guessing their username. This allows for heightened privacy for
users if they wish. The code works by adding the usernames of users why
opt-out to a text file and then reading from that same file to check if
that username exists in the file after a username is entered on the
match.html page.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>