Added categories for compatibility assessment
Now instead of just getting an overall compatibility score, the users
can choose a specific category that they want to see their score in.
There are specific question in each category and if both answers are the
same from either user within that category, the 'matches' value is
increased by two.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>