Made edits to the welcome page
On the welcome page I added gifs that I created on Canva which add a
nicer look to the website. These gifs were crafted from a template that
is freely available on Canva. The images added to them were obtained
through Pinterest. The first floral image: https://pin.it/3OzhJaPKp
Accessed 18 February, 2024. The watercolour people image:
https://pin.it/6fBOJLSvr Accessed 18 February, 2024.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>