The issues with original dataset are fixed
The original dataset had a few issues. It was not unambiguous, ordinal
or fine grained enough. This was due to the use of 'Indifferent' and
'Not Sure' as options within the survey and rows in the dataset. It is
not good to have neutral options in instances like this. I have stuck
with the 4-point scale however in the interest of time since gathering
survey data again would not be feasible at this point.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>