Email now sent to user with cluster result
During the semester one fyp demo, this was a suggested addition to make
to the application so that users are able to go back and check when they
last took the survey and to reflect on older categories they had been
placed in. So I have set up an app password for my google account and
that is used to send emails to the emails of the current logged in user
when they click the 'submit' button on the survey.html page.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>