Changed method of storing graphs
After gathering feedback of the application a common thing mentioned by
users was how long the graph page takes to load. After timing it, it was
taking roughly 13 seconds with the buffer method being used. I have now
opted for saving the graphs as images directly to the file system to
have the page load instantly and if any of the images are not present in
the file system, they are generated again. This removes the previous
bottleneck observed.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>