Fixed navbar elements showing and lots of styling
I adjusted the view of navbar elements showing since users could access
all pages even when survey was not completed which resulted in issues
and page contents not loading. I then made a lot of style changes,
mostly through the addition of images I had created on Canva, with the
use of some free templates as well as just some overall colour changes
which made a major improvement.

Signed-off-by: Saoirse O'Donovan <saoirseodonovan13@gmail.com>